fall back to a single fix ILVES implementation. update docs and examples.
