+++ examples/PACKAGES/ilves/README
@@ -1,47 +1,57 @@
 Examples for the ILVES bond/angle constraint solver.
 
-The ILVES solver comes in two variants:
+ILVES is invoked via "fix ilves".  The fix gathers the complete bond /
+angle topology onto every MPI rank at init via MPI_Allgatherv and
+builds a per-rank constraint list including every cluster intersecting
+the rank's local atoms.  This supports arbitrarily large connected
+constraint clusters (e.g. all C-C backbone bonds of a long polymer
+chain held in a single cluster) and any cluster topology -- including
+chains, rings, branched stars, or fully connected meshes.
 
-  * fix ilves/global -- gathers the complete bond/angle topology onto
-    every MPI rank at init.  Per-rank memory scales with the total
-    system size, but the variant handles arbitrarily large connected
-    constraint clusters (e.g. all C-C backbone bonds of a long polymer
-    chain in a single cluster).
-
-  * fix ilves/local  -- builds the constraint list from each rank's
-    own bond/angle storage only; no global topology gather, no
-    natoms-sized cluster table.  Every constraint cluster must fit
-    within a single subdomain plus the communication cutoff, and the
-    cluster topology must be star-shaped (one central atom with k
-    leaves) or a 1+1 pair.  Requires "newton off bond".  Suitable for
-    water HOH constraints, methyl / methane groups, isolated C-H bonds
-    -- the common ILVES use cases.
-
-Both variants accept the same selectors (b/a/t/m) and keywords
-(kbond, store, variant) as fix shake.  So the fix shake usage:
+Selectors and keywords mirror fix shake: the fix shake form
 
     fix cstr all shake 1.0e-4 25 0 b 4 6 8 10 12 14 18 a 31
 
-becomes either:
+becomes
 
-    fix cstr all ilves/global 1.0e-4 25 0 b 4 6 8 10 12 14 18 a 31
-    fix cstr all ilves/local  1.0e-4 25 0 b 4 6 8 10 12 14 18 a 31
+    fix cstr all ilves 1.0e-4 25 0 b 4 6 8 10 12 14 18 a 31
 
 Two solver variants are available via the *variant* keyword:
-  * variant = ilvesf (default): symmetric Jacobian using r_k.r_l in
-    off-diagonals; solved with in-place Cholesky.
-  * variant = ilves : the structurally-symmetric exact Newton
-    Jacobian using s_k.r_l in off-diagonals; solved with LU.
+  * variant = fast (default): symmetric Jacobian using r_k.r_l in
+    off-diagonals; solved with banded Cholesky on an RCM-permuted
+    matrix, and the factor is cached and reused across Newton
+    iterations within a step.  Up to ~5x faster than the full variant
+    on the all-backbone polymer / protein benchmark.
+  * variant = full: the structurally-symmetric exact Newton Jacobian
+    using s_k.r_l in off-diagonals; solved with LU.  Required when
+    the constraint geometry has degenerate clusters that the
+    Cholesky path cannot factor (the solver will fall back to LU per
+    cluster as needed, but starting in "full" mode skips the
+    fallback overhead).
 
-Both converge to the exact solution of the constraint equations.
+Both variants converge to the exact solution of the constraint
+equations.
+
+For near-linear angle constraints (theta_0 close to 180 deg, common
+in coarse-grain polymer backbones), use the *linearangle* keyword to
+switch the angle's constraint topology from {AB, BC, AC} (which is
+rank-deficient at exactly 180 deg) to {AB, AC, B-M} where M is the
+midpoint of the A-C virtual bond:
+
+    fix cstr all ilves 1.0e-5 30 0 b 1 a 1 variant fast linearangle 165
+
+Default threshold is 165 deg; set to 180 to disable.
 
 Example input files:
-  in.peptide-ilves    -- peptide + water with ilves/global, exact-Newton variant
-  in.peptide-ilvesf   -- peptide + water with ilves/global, symmetric variant
-  in.peptide-local    -- peptide + water with ilves/local (newton off bond)
-  in.peptide-mix      -- peptide groups via ilves/global, water via fix shake
-  in.peptide-allbonds -- ilves/global constraining ALL bonds, demonstrating
-                         arbitrarily-large cluster support (not feasible with
-                         ilves/local, since the backbone forms a multi-hop chain)
+  in.peptide-ilves    -- peptide + water with ilves, exact-Newton variant (full)
+  in.peptide-ilvesf   -- peptide + water with ilves, symmetric variant (fast)
+  in.peptide-mix      -- peptide groups via ilves, water via fix shake
+  in.peptide-allbonds -- ilves constraining ALL bonds, demonstrating
+                         arbitrarily-large cluster support
   in.peptide-shake    -- baseline using fix shake for comparison
   in.peptide-rattle   -- baseline using fix rattle for comparison
+  in.rhodo-ilves      -- rhodopsin with C-H bonds + H-O-H water angle (full)
+  in.rhodo-ilvesf     -- rhodopsin with same constraints (fast)
+  in.rhodo-allbonds   -- rhodopsin with all bonds + water angle
+  in.bulk_co2         -- bulk CO2 fluid (rigid linear; see comments for
+                         caveats on exactly-180 deg symmetric angles)
