+++ examples/PACKAGES/ilves/in.bulk_co2
@@ -1,0 +1,86 @@
+# Bulk CO2 molecular fluid, fix ilves
+# Rigid CO2 TraPPE model
+# [Potoff and J.I. Siepmann, Vapor-liquid equilibria of
+# mixtures containing alkanes, carbon dioxide and
+# nitrogen AIChE J., 47,1676-1682 (2001)].
+#
+# Note: a near-linear (or exactly linear) triatomic like CO2 is a
+# challenging case for any iterative constraint solver because the
+# Jacobian is rank-deficient at theta = 180 deg.  We use the
+# linearangle handling and a moderately large |B-M| clamp (Lmin=0.1,
+# corresponding to ~9 deg off linear), plus an initial minimize, to
+# keep the solver stable.  For production simulations of truly rigid
+# linear molecules, fix rigid/small is the recommended approach.
+
+# variables available on command line
+
+variable        temp index 338.0
+variable        lbox index 25.0
+variable        spacing index 5.0
+
+# global model settings
+
+units           real
+atom_style      full
+boundary        p p p
+pair_style      lj/cut/coul/long  14
+pair_modify     mix arithmetic tail yes
+kspace_style    pppm 0.00001
+bond_style      harmonic
+angle_style     harmonic
+
+# box, start molecules on simple cubic lattice
+
+lattice         sc ${spacing}
+region          box block 0 ${lbox} 0 ${lbox} 0 ${lbox} units box
+create_box      2 box                       &
+                bond/types 1                &
+                angle/types 1               &
+                extra/bond/per/atom 2       &
+                extra/angle/per/atom 1      &
+                extra/special/per/atom 2
+molecule        co2mol CO2.txt
+create_atoms    0 box mol co2mol 464563 units box
+
+# rigid CO2 TraPPE model
+
+pair_coeff      1   1  0.053649   2.8
+pair_coeff      2   2  0.156973   3.05
+bond_coeff      1       0       1.16
+angle_coeff     1       0       180
+
+# masses
+
+mass 1 12.0107
+mass 2 15.9994
+
+# MD settings
+
+group           co2 type 1 2
+neighbor        2.0 bin
+neigh_modify    every 1 delay 2 check yes
+velocity        all create ${temp} 54654
+timestep        1.0
+
+# rigid constraints with thermostat
+
+fix             rigid all ilves 0.00001 25 100 t 1 2 a 1 variant fast linearangle 165 0.1
+fix             move  all nvt temp ${temp} ${temp} 100.0
+
+thermo_style    custom step temp press pe ke density
+thermo          100
+
+# minimize first to relax initial lattice overlap.  Without this, the
+# random-lattice initial positions put O atoms only ~2.2 Angstrom
+# apart between neighbouring CO2 molecules (well inside the LJ sigma
+# of 3.05 Angstrom), and the resulting initial forces produce an
+# unconstrained step large enough to blow up the constraint solver.
+minimize 0.0 0.0 10000 100000
+
+reset_timestep 0
+
+dump viz all image 10 myimage-*.png element type size 800 800 zoom 1.21 shiny 0.6 fsaa yes bond atom type view 60 30 box yes 0.025 subbox no 0.0 axes no 0.0 0.0 center s 0.5 0.5 0.5
+dump_modify viz pad 9 color red 1 0 0 color forestgreen 0.133333 0.545098 0.133333 color blue 0 0 1 color gold 1 0.843137 0 color cyan 0 1 1 color magenta 1 0 1 color silver 0.431373 0.431373 0.431373 color orange 1 0.501961 0 color lime 0 1 0 color gray 0.501961 0.501961 0.501961 color darkred 0.545098 0 0 color darkgreen 0 0.392157 0 color darkblue 0 0 0.545098 color darkcyan 0 0.545098 0.545098 color darkmagenta 0.545098 0 0.545098 color darkgray 0.270588 0.270588 0.270588 acolor 1 red acolor 2 forestgreen boxcolor yellow backcolor black backcolor2 white axestrans 1 boxtrans 1 atrans * 1 lights 0 0.9 0.45 0.9 element C O adiam 1 0.85 adiam 2 0.76 color map1 0.000 0.227 0.427 color map2 0.459 0.055 0.075 amap min max cf 0.0 3 min map1 0.5 white max map2
+
+run             2000
+
