+++ examples/PACKAGES/ilves/in.peptide-allbonds
@@ -26,6 +26,6 @@ fix             1 all nvt temp 275.0 275.0 100.0 tchain 1
 group           peptide type <= 12
 group           waters  type > 12
 
-fix             cstr all ilves/global 0.00001 20 100 b 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 a 31 variant ilvesf
+fix             cstr all ilves 0.00001 20 100 b 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 a 31 variant fast
 
 run             300
