--- examples/PACKAGES/ilves/in.peptide-local
@@ -1,37 +1,0 @@
-# Solvated 5-mer peptide -- ILVES constraint solver, local variant
-# (mirror of in.peptide-ilvesf using fix ilves/local; constrains only
-#  water bonds + the HOH angle, which form 3-atom star clusters)
-
-units           real
-newton          on off    # newton bond OFF is required by fix ilves/local
-atom_style      full
-
-pair_style      lj/charmm/coul/long 8.0 10.0 10.0
-bond_style      harmonic
-angle_style     charmm
-dihedral_style  charmm
-improper_style  harmonic
-kspace_style    pppm 0.0001
-
-read_data       data.peptide
-
-neighbor        2.0 bin
-neigh_modify    delay 5
-
-timestep        2.0
-
-thermo_style    multi
-thermo          50
-
-fix             1 all nvt temp 275.0 275.0 100.0 tchain 1
-
-group           peptide type <= 12
-group           waters  type > 12
-
-# fix ilves/local: constraint clusters must fit within a subdomain plus
-# the communication cutoff.  Water HOH (b 18 + a 31) gives 3-atom star
-# clusters that comfortably fit.  Peptide C-H bonds (b 4 6 8 10 12 14)
-# form 1+k stars centered on each heavy atom and similarly fit.
-fix             cstr all ilves/local 0.00001 20 100 b 4 6 8 10 12 14 18 a 31 variant ilvesf
-
-run             300
