+++ examples/PACKAGES/ilves/in.peptide-mix
@@ -26,7 +26,7 @@ fix             1 all nvt temp 275.0 275.0 100.0 tchain 1
 group           peptide type <= 12
 group           waters  type > 12
 
-fix             s_rigid peptide ilves/global 0.00001 20 100 b 4 6 8 10 12 14
+fix             s_rigid peptide ilves 0.00001 20 100 b 4 6 8 10 12 14
 fix             w_rigid waters shake 0.00001 20 100 b 18 a 31
 
 run             300
